Add validate.InRange

+++ staging/src/k8s.io/apimachinery/pkg/util/validation/validation_test.go
@@ -158,9 +158,9 @@ func TestIsValidPortNum(t *testing.T) {
 
 func TestIsInRange(t *testing.T) {
 	goodValues := []struct {
-		value int
-		min   int
-		max   int
+		value int64
+		min   int64
+		max   int64
 	}{{1, 0, 10}, {5, 5, 20}, {25, 10, 25}}
 	for _, val := range goodValues {
 		if msgs := IsInRange(val.value, val.min, val.max); len(msgs) > 0 {
@@ -169,9 +169,9 @@ func TestIsInRange(t *testing.T) {
 	}
 
 	badValues := []struct {
-		value int
-		min   int
-		max   int
+		value int64
+		min   int64
+		max   int64
 	}{{1, 2, 10}, {5, -4, 2}, {25, 100, 120}}
 	for _, val := range badValues {
 		if msgs := IsInRange(val.value, val.min, val.max); len(msgs) == 0 {
